+++ client/mapView/MapViewController.cpp
@@ -154,6 +154,7 @@ void MapViewController::updateBefore(uint32_t timeDelta)
 		adventureContext->settingShowGrid = settings["gameTweaks"]["showGrid"].Bool();
 		adventureContext->settingShowVisitable = settings["session"]["showVisitable"].Bool();
 		adventureContext->settingShowBlocked = settings["session"]["showBlocked"].Bool();
+		adventureContext->settingSpellRange = settings["session"]["showSpellRange"].Bool();
 	}
 }